dev: [Quinn] moved header line to common folder
